refactor : move components to global folder
